Renamed map component
